Rewrite the code that creates the section headings
The problem of the previous method was the checks were done during the
creation of the glossary entries files, but the entries are not sorted
in the source files.

The new method reads each glossary entry file in order from the
temporary directory, and creates additional files with the heading
markup.

FossilOrigin-Name: b401494516e19d34f70ce78490df01dd93e0d623ff5811048f0ec215f0d28728