Change the escaping of the backslash to "&#92;"
`:backslash:` is a predefined attribute in Asciidoctor, and that's why
the backslash was escaped as `{backslash}` by Glosara. But the value of
`:backslash:` is the actual character (ASCII 92), which still causes
problems to the renderers, while the numeric character reference "&#92;"
works fine.

`:backslash:` could be reassigned to "&#92;" in the source document, but
it's simpler to avoid that requirement and do the replacement in
Glosara.

FossilOrigin-Name: 666df107502dbe5b021630b3f4850c5b970bc029fea65102c51a34be399f44d8