Fix the position of inline unique ids in titles
In order to be recognized by Asciidoctor as the unique id of a title,
not just as an anchor, the markup must be at the end of the title, after
a space. Therefore the unique id has been moved there, and the common
anchor for homonyms is assoctiated to an empty block.

This change avoids the problem that the cross references didn't work in
the documents made by Asciidoctor EPUB3.

See: https://github.com/asciidoctor/asciidoctor/issues/3633.

FossilOrigin-Name: 4ee1bfb9c7667efbf553c4546d5c40681b939af5bfb9998872f8cbee7e54457d